-cuda integral (partial merge fenix_realisation) - update start.sh

+++ cuda_integral/include/math/integration_args.h
@@ -12,9 +12,8 @@ struct integration_args {
     point start = point{0.0, 0.0};
     point end = point{0.0, 0.0};
     const f_params conf{};
-    const uint16_t flow_n = 1;
+    const ptrdiff_t flow_n = 1;
     double dx = 0.0, dy = 0.0;
-//    const size_t steps;
 };
 
 #endif //LAB_2_PARALLEL_INTEGRATION_INTEGRATION_ARGS_H
